# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BirrStream API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a daily task and earn reward
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompleteResult"
        "400":
          description: Already completed or error
...
  /withdrawals:
    post:
      operationId: submitWithdrawal
      tags: [withdrawals]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          description: Below reserve limit or insufficient balance
...
  /withdrawals/settings:
    get:
      operationId: getWithdrawalSettings
      tags: [withdrawals]
      summary: Get user's bank/wallet withdrawal settings
      responses:
        "200":
          description: Withdrawal settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalSettings"
    put:
      operationId: updateWithdrawalSettings
      tags: [withdrawals]
      summary: Update bank/wallet withdrawal settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalSettingsInput"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalSettings"
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/delete:
    delete:
      operationId: deleteAccount
      tags: [user]
      summary: Delete user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountInput"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Confirmation string mismatch
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        mainBalance:
          type: number
        totalYield:
          type: number
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
        referralCode:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - fullName
        - username
        - email
        - mainBalance
        - totalYield
        - totalDeposited
        - totalWithdrawn
        - referralCode
        - isAdmin
        - createdAt
...
    TaskCompleteResult:
      type: object
      properties:
        success:
          type: boolean
        rewardEarned:
          type: number
        newBalance:
          type: number
        message:
          type: string
      required:
        - success
        - rewardEarned
        - newBalance
        - message
...
    Withdrawal:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        bankName:
          type: string
        accountName:
          type: string
        walletId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - amount
        - status
        - bankName
        - accountName
        - walletId
        - createdAt
    WithdrawalInput:
      type: object
      properties:
        amount:
          type: number
      required:
        - amount
    WithdrawalSettings:
      type: object
      properties:
        bankName:
          type: ["string", "null"]
        accountName:
          type: ["string", "null"]
        walletId:
          type: ["string", "null"]
        isConfigured:
          type: boolean
      required:
        - isConfigured
    WithdrawalSettingsInput:
      type: object
      properties:
        bankName:
          type: string
        accountName:
          type: string
        walletId:
          type: string
      required:
        - bankName
        - accountName
        - walletId
...
    UserProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
    DeleteAccountInput:
      type: object
      properties:
        confirmationText:
          type: string
      required:
        - confirmationText